+++ index.html
@@ -12,7 +12,7 @@
 <body>
     <header>
         <nav>
-            <h3>Awesome books</h3>
+            <h3>awesome books</h3>
             <ul>
                 <li><a href="#" onclick="a.lists()">Lists</a></li>
                 <li><a href="#" onclick="a.addNewBook()">Add new</a> </li>
@@ -33,7 +33,7 @@
         <form action="">
             <input type="text" id="title" placeholder="title">
             <input type="text" id="author" placeholder="author"> 
-            <div id="formbut"><button id="addbtn"> Add</button></div>
+            <div id="formbut"><button id="addbtn" onclick="a.addBooks()"> Add</button></div>
         </form>
        </div>  
     </section>
@@ -60,5 +60,6 @@
  </footer>
     
     <script type="module" src="./index.js"></script>
+    <script src='https://cdnjs.cloudflare.com/ajax/libs/luxon/3.3.0/luxon.min.js'></script>
 </body>
 </html>
