+++ modules/datetime.js
@@ -1,0 +1,6 @@
+const dt = luxon.DateTime.now().toFormat('MMMM d yyyy, HH:mm:ss') + luxon.DateTime.now().toFormat(' a').toLowerCase();
+
+export default dt;
+
+console.log(dt); 
+
